+++ generators/typescript/utils/commons/src/core-utilities/CoreUtilities.ts
@@ -6,6 +6,7 @@ import { FormDataUtils } from "./form-data-utils/FormDataUtils";
 import { Pagination } from "./pagination/Pagination";
 import { Runtime } from "./runtime/Runtime";
 import { StreamUtils } from "./stream-utils/StreamUtils";
+import { Utils } from "./utils/Utils";
 import { Zurg } from "./zurg/Zurg";
 
 export interface CoreUtilities {
@@ -18,4 +19,5 @@ export interface CoreUtilities {
     formDataUtils: FormDataUtils;
     runtime: Runtime;
     pagination: Pagination;
+    utils: Utils;
 }