+++ generators/typescript/utils/commons/src/core-utilities/utils/Utils.ts
@@ -1,0 +1,7 @@
+import { ts } from "ts-morph";
+
+export interface Utils {
+    readonly setObjectProperty: {
+        _invoke: (args: { referenceToObject: ts.Expression; path: string; value: ts.Expression }) => ts.Expression;
+    };
+}