+++ seed/ts-express/pagination/api/resources/users/service/UsersService.ts
@@ -27,6 +27,20 @@ export interface UsersServiceMethods {
         },
         next: express.NextFunction
     ): void | Promise<void>;
+    listWithBodyCursorPagination(
+        req: express.Request<
+            never,
+            SeedPagination.ListUsersPaginationResponse,
+            SeedPagination.ListUsersBodyCursorPaginationRequest,
+            never
+        >,
+        res: {
+            send: (responseBody: SeedPagination.ListUsersPaginationResponse) => Promise<void>;
+            cookie: (cookie: string, value: string, options?: express.CookieOptions) => void;
+            locals: any;
+        },
+        next: express.NextFunction
+    ): void | Promise<void>;
     listWithOffsetPagination(
         req: express.Request<
             never,
@@ -38,6 +52,20 @@ export interface UsersServiceMethods {
                 order?: SeedPagination.Order;
                 starting_after?: string;
             }
+        >,
+        res: {
+            send: (responseBody: SeedPagination.ListUsersPaginationResponse) => Promise<void>;
+            cookie: (cookie: string, value: string, options?: express.CookieOptions) => void;
+            locals: any;
+        },
+        next: express.NextFunction
+    ): void | Promise<void>;
+    listWithBodyOffsetPagination(
+        req: express.Request<
+            never,
+            SeedPagination.ListUsersPaginationResponse,
+            SeedPagination.ListUsersBodyOffsetPaginationRequest,
+            never
         >,
         res: {
             send: (responseBody: SeedPagination.ListUsersPaginationResponse) => Promise<void>;
@@ -164,6 +192,49 @@ export class UsersService {
                 next(error);
             }
         });
+        this.router.post("", async (req, res, next) => {
+            const request = serializers.ListUsersBodyCursorPaginationRequest.parse(req.body);
+            if (request.ok) {
+                req.body = request.value;
+                try {
+                    await this.methods.listWithBodyCursorPagination(
+                        req as any,
+                        {
+                            send: async (responseBody) => {
+                                res.json(
+                                    serializers.ListUsersPaginationResponse.jsonOrThrow(responseBody, {
+                                        unrecognizedObjectKeys: "strip",
+                                    })
+                                );
+                            },
+                            cookie: res.cookie.bind(res),
+                            locals: res.locals,
+                        },
+                        next
+                    );
+                    next();
+                } catch (error) {
+                    if (error instanceof errors.SeedPaginationError) {
+                        console.warn(
+                            `Endpoint 'listWithBodyCursorPagination' unexpectedly threw ${error.constructor.name}.` +
+                                ` If this was intentional, please add ${error.constructor.name} to` +
+                                " the endpoint's errors list in your Fern Definition."
+                        );
+                        await error.send(res);
+                    } else {
+                        res.status(500).json("Internal Server Error");
+                    }
+                    next(error);
+                }
+            } else {
+                res.status(422).json({
+                    errors: request.errors.map(
+                        (error) => ["request", ...error.path].join(" -> ") + ": " + error.message
+                    ),
+                });
+                next(request.errors);
+            }
+        });
         this.router.get("", async (req, res, next) => {
             try {
                 await this.methods.listWithOffsetPagination(
@@ -194,6 +265,49 @@ export class UsersService {
                     res.status(500).json("Internal Server Error");
                 }
                 next(error);
+            }
+        });
+        this.router.post("", async (req, res, next) => {
+            const request = serializers.ListUsersBodyOffsetPaginationRequest.parse(req.body);
+            if (request.ok) {
+                req.body = request.value;
+                try {
+                    await this.methods.listWithBodyOffsetPagination(
+                        req as any,
+                        {
+                            send: async (responseBody) => {
+                                res.json(
+                                    serializers.ListUsersPaginationResponse.jsonOrThrow(responseBody, {
+                                        unrecognizedObjectKeys: "strip",
+                                    })
+                                );
+                            },
+                            cookie: res.cookie.bind(res),
+                            locals: res.locals,
+                        },
+                        next
+                    );
+                    next();
+                } catch (error) {
+                    if (error instanceof errors.SeedPaginationError) {
+                        console.warn(
+                            `Endpoint 'listWithBodyOffsetPagination' unexpectedly threw ${error.constructor.name}.` +
+                                ` If this was intentional, please add ${error.constructor.name} to` +
+                                " the endpoint's errors list in your Fern Definition."
+                        );
+                        await error.send(res);
+                    } else {
+                        res.status(500).json("Internal Server Error");
+                    }
+                    next(error);
+                }
+            } else {
+                res.status(422).json({
+                    errors: request.errors.map(
+                        (error) => ["request", ...error.path].join(" -> ") + ": " + error.message
+                    ),
+                });
+                next(request.errors);
             }
         });
         this.router.get("", async (req, res, next) => {